docs: 子项目 CLAUDE.md 与 TASK→TODO；编译器与工具链更新
- 为 compiler/tools 各 crate 补充可操作的 CLAUDE.md（模块、API、流水线衔接）
- 移除各 crate TASK.md，新增对应 TODO.md；删除根目录 PLAN.md、Ralph.md
- x-codegen-asm、多后端、HIR/MIR/LIR、lexer/parser、typechecker、x-cli、x-lsp、x-syntax-gen 等实现与修复
- library/stdlib 增加 README

Made-with: Cursor

+++ tools/x-lsp/src/analysis/symbol.rs
@@ -86,8 +86,8 @@ impl SymbolTable {
 
         // Find symbol that contains this position
         for symbol in &self.symbols {
-            let start = symbol.span.start as usize;
-            let end = symbol.span.end as usize;
+            let start = symbol.span.start;
+            let end = symbol.span.end;
             if byte_offset >= start && byte_offset <= end {
                 return Some(symbol);
             }
